feat(sealed-types): introduce `sealed` modifier for classes, structs, and enums
- Added support for the `sealed` modifier, preventing user-defined `extend` blocks from targeting sealed types, while allowing compiler-synthesized extends.
- Updated the parser to recognize `sealed` as a valid modifier for class, struct, and enum declarations.
- Enhanced the semantic analyzer to enforce restrictions on extending sealed types, with appropriate diagnostics for violations.
- Introduced tests to validate the behavior of sealed types, ensuring that attempts to extend them are correctly rejected.
- Updated documentation to reflect the new `sealed` modifier and its implications for type extension.

+++ src/execution/host/datetime.rs
@@ -30,8 +30,8 @@ pub fn link_datetime_functions(linker: &mut Linker<()>) -> Result<()> {
     linker.func_wrap("Dream", "dateLocalOffsetMinutes", |millis: i64| -> i32 {
         use chrono::{Local, TimeZone};
         match Local.timestamp_millis_opt(millis) {
-            chrono::LocalResult::Single(dt) => (dt.offset().local_minus_utc() / 60) as i32,
-            chrono::LocalResult::Ambiguous(dt, _) => (dt.offset().local_minus_utc() / 60) as i32,
+            chrono::LocalResult::Single(dt) => dt.offset().local_minus_utc() / 60,
+            chrono::LocalResult::Ambiguous(dt, _) => dt.offset().local_minus_utc() / 60,
             chrono::LocalResult::None => 0,
         }
     })?;
